hammmerhead: Fix rear camera not working at all until next reboot Something goes wrong when the cameras are probed in the kernel, causing them to probe successfully but the rear camera will not work after that (VIDIOC_S_CTRL ioctl fails). The front camera, however, works fine all the time.
Restarting the camera server fixes the issue. This issue probably has
something to do with the fact that the front camera is always probed
first, but changing the probe order causes non-ZSL capture to break, so
this is the only alternative left (fixing non-ZSL capture with the correct
probe order would require proprietary camera modifications).

Change-Id: Ia2eea4a58ff0b3e01a52396add12b200f48eb1c4
Signed-off-by: Marco Marinho <marcomarinho12@gmail.com>