hammerhead: disable CPU retention
There is a bug in the chipset where the attempt to change the state of a
clock happens during the retention phase on the target cpu. Krait does
not like this and dies. This patch should fix that issue.

This is specific to the msm8974 revision of the krait 400 chipset, and
not the msm8974pro revision.

Change-Id: I9bda0cd0801dce6e128bd5046fe60e751efa7953
Signed-off-by: Marco Marinho <marcomarinho12@gmail.com>